krowdkam/requirements.txt
asgiref==3.4.1
Django==3.2.5
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
opencv-python==4.5.2.54
djangorestframework==3.12.4
djangorestframework-simplejwt
torch==1.9.0
h5py
scipy
Pillow==8.3.1
matplotlib
imutils
pywhatkit